Move GlobalVariableInfo class to its own files
This seems to fix the bug compiling with GCC 13.2 described in #1425.
