+++ include/natalie/global_env.hpp
@@ -5,6 +5,7 @@
 #include "natalie/encodings.hpp"
 #include "natalie/forward.hpp"
 #include "natalie/gc.hpp"
+#include "natalie/global_variable_info.hpp"
 #include "natalie/method_missing_reason.hpp"
 #include "tm/hashmap.hpp"
 
@@ -16,20 +17,6 @@ extern "C" {
 
 class GlobalEnv : public Cell {
 public:
-    class GlobalVariableInfo : public Cell {
-    public:
-        GlobalVariableInfo(Object *object)
-            : m_object { object } { }
-
-        void set_object(Object *object) { m_object = object; }
-        Object *object() { return m_object; }
-
-        virtual void visit_children(Visitor &visitor) override final;
-
-    private:
-        class Object *m_object { nullptr };
-    };
-
     static GlobalEnv *the() {
         if (s_instance)
             return s_instance;
