+++ src/global_variable_info.cpp
@@ -1,0 +1,9 @@
+#include "natalie.hpp"
+
+namespace Natalie {
+
+void GlobalVariableInfo::visit_children(Visitor &visitor) {
+    visitor.visit(m_object);
+}
+
+}
